+++ src/reducers/__tests__/sourcereel.ts
@@ -2,19 +2,13 @@ import {
   IAction,
   RECORD_EDITOR_INIT_VALUE,
   RECORD_INPUT,
-  SAVE_SOURCECAST_DATA,
   TIMER_PAUSE,
   TIMER_RESET,
   TIMER_RESUME,
   TIMER_START,
   TIMER_STOP
 } from '../../actions/actionTypes';
-import {
-  ICodeDelta,
-  Input,
-  IPlaybackData,
-  RecordingStatus
-} from '../../components/sourcecast/sourcecastShape';
+import { ICodeDelta, Input, RecordingStatus } from '../../components/sourcecast/sourcecastShape';
 import { reducer } from '../sourcereel';
 import { defaultWorkspaceManager } from '../states';
 
@@ -72,48 +66,6 @@ describe('RECORD_INPUT', () => {
         ...defaultWorkspaceManager.sourcereel.playbackData,
         inputs: [...defaultWorkspaceManager.sourcereel.playbackData.inputs, input]
       }
-    });
-  });
-});
-
-describe('SAVE_SOURCECAST_DATA', () => {
-  test('saves sourcecastData correctly', () => {
-    const codeDelta: ICodeDelta = {
-      start: {
-        row: 0,
-        column: 1
-      },
-      end: {
-        row: 0,
-        column: 2
-      },
-      action: 'insert',
-      lines: ['a']
-    };
-    const input: Input = {
-      time: 1,
-      type: 'codeDelta',
-      data: codeDelta
-    };
-    const playbackData: IPlaybackData = {
-      init: {
-        editorValue: ''
-      },
-      inputs: [input]
-    };
-
-    const payload = {
-      title: 'Test Title',
-      description: 'Test Description',
-      audioUrl: 'someUrl.com',
-      playbackData
-    };
-
-    const action: IAction = generateAction(SAVE_SOURCECAST_DATA, payload);
-    const result = reducer(defaultWorkspaceManager.sourcereel, action);
-    expect(result).toEqual({
-      ...defaultWorkspaceManager.sourcereel,
-      ...payload
     });
   });
 });
